requirements.txt
psycopg2-binary
matplotlib
pandas
findspark